+++ ci/docker-compose-local-readwrite.yml
@@ -16,13 +16,26 @@ services:
       - localstack
 
   db:
-    image: postgres:15
+    image: postgres:11
     restart: unless-stopped
     ports:
       - 5432:5432
     environment:
       - POSTGRES_PASSWORD=LocalPassword
       - POSTGRES_DB=notification_api
+    volumes:
+      - ./db-data:/var/db-data
+
+  db-read:
+    image: postgres:11
+    restart: unless-stopped
+    ports:
+      - 5433:5432
+    environment:
+      - POSTGRES_PASSWORD=LocalPassword
+      - POSTGRES_DB=notification_api
+    volumes:
+      - ./db-data-read:/var/db-data-read
 
   migrations:
     image: notification_api
